Add files via upload
The core concept of the LRU algorithm is to evict the oldest data from the cache to accommodate more data. To implement an LRU cache we use two data structures: a hashmap and a doubly linked list. A doubly linked list helps in maintaining the eviction order and a hashmap helps with O(1) lookup of cached keys.
he idea is very basic, i.e. keep inserting the elements at the head.

if the element is not present in the list then add it to the head of the list
if the element is present in the list then move the element to the head and shift the remaining element of the list
Note that the priority of the node will depend upon the distance of that node from the Head, the closest the node is to the head, higher the priority it has. So when the Cache size is full and we need to remove some element, we remove the element adjacent to the tail of the doubly linked list.